requirements.txt
spotipy>=2.23.0
yt-dlp>=2024.10.7
mutagen>=1.47.0
librosa>=0.10.2
numpy>=1.24.0
scipy>=1.11.0
python-dotenv>=1.0.1
tqdm>=4.66.0
